% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{add_metrics}
\alias{add_metrics}
\title{Add selected network metrics to actors}
\usage{
add_metrics(profiles, net)
}
\arguments{
\item{profiles}{tibble with list of actors in net}

\item{net}{tibble with two columns, two columns: \code{actor_handle}
(the Blue Sky Social actor who is followING another) and \code{follows_handle}
(the actor being followed).}
}
\value{
tibble with profiles and extra columns for metrics
}
\description{
Add selected network metrics to actors
}
