% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueskynet-package.R
\docType{package}
\name{blueskynet-package}
\alias{blueskynet}
\alias{blueskynet-package}
\title{blueskynet: Analyze Networks on then Bluesky Social Network}
\description{
Based on a given actor (user) on Bluesky, this will generate a network based on what users the actor followes. The output will be a graph with information on who follows who among this set of actors. Also included are a few tools from other packages for network analysis.
}
\author{
\strong{Maintainer}: Lasse Hjorth Madsen \email{lassehjorthmadsen@gmail.com}

}
\keyword{internal}
