% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{com_labels}
\alias{com_labels}
\title{Label communities}
\usage{
com_labels(profiles, top = 3)
}
\arguments{
\item{profiles}{tibble with profile information, assumed to contain
a column called "community" for grouping, and a column called
"description" that contain the text to analyze}

\item{top}{number of words to use in label, defaults to 3}
}
\value{
tibble with group id and label
}
\description{
Label communities
}
