% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{extract_follow_subjects}
\alias{extract_follow_subjects}
\title{Extract subject DIDs from follow records}
\usage{
extract_follow_subjects(follow_records)
}
\arguments{
\item{follow_records}{list, output from get_all_follow_records()}
}
\value{
character vector of DIDs
}
\description{
Extract subject DIDs from follow records
}
