% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{find_follow_record}
\alias{find_follow_record}
\title{Find follow record for specific actor}
\usage{
find_follow_record(follow_records, actor_did)
}
\arguments{
\item{follow_records}{list, output from get_all_follow_records()}

\item{actor_did}{character, DID to search for}
}
\value{
single follow record or NULL if not found
}
\description{
Find follow record for specific actor
}
