% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_followers}
\alias{get_followers}
\title{Get followers for an actor}
\usage{
get_followers(actor, token, max_retries = 3, retry_delay = 5)
}
\arguments{
\item{actor}{account identifier}

\item{token}{token for api}

\item{max_retries}{number of times to retry on failure (default 3)}

\item{retry_delay}{delay in seconds between retries (default 5)}
}
\value{
tibble with response information
}
\description{
Get followers for an actor
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/graph/getFollowers.json}{lexicons/app/bsky/graph/getFollowers.json}
}

