% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_profiles_error}
\alias{get_profiles_error}
\title{Get profile for an actor/actors}
\usage{
get_profiles_error(
  actors,
  token,
  refresh_tok,
  chunksize = 25,
  refresh_interval = 20
)
}
\arguments{
\item{actors}{character, actor handle}

\item{token}{character, api token}

\item{refresh_tok}{character, refresh token}

\item{chunksize}{integer, the number of actors per request;
defaults to 25, currently the maximum number allowed}

\item{refresh_interval}{integer, number of chunks to process before refreshing token}
}
\value{
tibble with profiles
}
\description{
Get profile for an actor/actors
}
