% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_profiles_old}
\alias{get_profiles_old}
\title{Get profile for an actor/actors}
\usage{
get_profiles_old(actors, token, chunksize = 25)
}
\arguments{
\item{actors}{character, actor handle}

\item{token}{character, api token}

\item{chunksize}{integer, the number of actors per request;
defaults to 25, currently the maximum number allowed}
}
\value{
tibble with profiles
}
\description{
Get profile for an actor/actors
}
