% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{unfollow_actor}
\alias{unfollow_actor}
\title{Unfollow an actor using record key}
\usage{
unfollow_actor(my_did, rkey, token)
}
\arguments{
\item{my_did}{character, did-identification of the actor that wants to unfollow someone}

\item{rkey}{character, record key of the follow relationship to delete}

\item{token}{character, api token}
}
\value{
response object
}
\description{
Unfollow an actor using record key
}
