% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{verify_token}
\alias{verify_token}
\title{Verify token}
\usage{
verify_token(token)
}
\arguments{
\item{token, }{character, api token}
}
\value{
boolean, is the token valid (TRUE) or not (FALSE)?
}
\description{
Verify token
}
