% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{build_network}
\alias{build_network}
\title{Build a complete network and helpful artifacts}
\usage{
build_network(key_actor, keywords, token, refresh_tok, threshold, ...)
}
\arguments{
\item{key_actor}{account identifier}

\item{keywords}{character vector with keywords that we check for in actors' description}

\item{token}{character, token for Blue Sky Social API}

\item{refresh_tok}{character, refresh token for Blue Sky Social API}

\item{threshold}{the threshold for including actors in the expansion: How many
followers must a prospect have, to be considered? If smaller than 1, threshold is used
as a proportion: How big a proportion must a prospect have, to be considered?}

\item{...}{Parameters passed on to \code{expand_net()}}
}
\value{
list of objects generated from \code{expand_net()}, \code{trim_net()}, \code{get_profiles()},
\code{create_widget()}, and \code{word_freqs()}
}
\description{
Build a complete network and helpful artifacts
}
