% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{com_labels}
\alias{com_labels}
\title{Label communities}
\usage{
com_labels(profiles, group = "community", text_field = "description", top = 3)
}
\arguments{
\item{profiles}{tibble with profile information}

\item{group}{field that contains grouping information, defaults to "community"}

\item{text_field}{field that contains text to use for word frequencies, defaults to "description"}

\item{top}{number of words to use in label, defaults to 3}
}
\value{
tibble with group id and label
}
\description{
Label communities
}
