% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{expand_net}
\alias{expand_net}
\title{Expand network}
\usage{
expand_net(
  net,
  keywords,
  token,
  refresh_tok,
  save_net = FALSE,
  file_name,
  threshold,
  max_iterations,
  sample_size = Inf
)
}
\arguments{
\item{net}{tibble with network connections (edges). Assumed to contain two columns:
\code{actor_handle} (the Blue Sky Social actor who is followING another) and \code{follows_handle}
(the actor being followed).}

\item{keywords}{character vector with keywords that we check for in actors' description
to determine if they belong in the expanded network}

\item{token}{character, token for Blue Sky Social API}

\item{refresh_tok}{character, refresh token for Blue Sky Social API}

\item{save_net}{boolean, should the net be saved incrementally as it is expanded? Useful
for not losing work in case of unexpected interruptions. Defaults to \code{FALSE}}

\item{file_name}{character, file name of the saved net. Defaults to 'dev/net/bignet_TIMESTAMP.rds'}

\item{threshold}{numeric, the threshold for including actors in the expansion: How many
followers must a prospect have, to be considered? If less then 1, interpreted as the \emph{fraction}
of the current net prospects must be followed by, to be considered. Defaults to 30.}

\item{max_iterations}{integer, the maximum iterations in the network expansion. Defaults to 50}

\item{sample_size}{integer, if we want to expand only with a sample of the prospects;
can be useful for testing purposes. Defaults to `Inf´, e.g. every worthy prospect is included
in the expansion.}
}
\value{
list with a) a tibble with the expanded network: \code{actor_handle} contains all network member,
\code{follows_handle} contain all relevant actors that they follow (but not necessarily meeting
the \code{threshold}.) b) a new token, c) a new refresh token -- since the tokens might have beed
refreshed while underway
}
\description{
Expand network
}
