% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{follow_actor}
\alias{follow_actor}
\title{Follow an actor or actors}
\usage{
follow_actor(my_did, actor_did, token)
}
\arguments{
\item{my_did}{character, did-identification of the actor that wants to follow someone}

\item{actor_did}{character, did-identification of the actor to be followed}

\item{token}{token for api}
}
\value{
response object
}
\description{
Follow an actor or actors
}
