% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_all_follow_records}
\alias{get_all_follow_records}
\title{Get all follow records from user's repository}
\usage{
get_all_follow_records(my_did, token, max_retries = 3, retry_delay = 5)
}
\arguments{
\item{my_did}{character, did-identification of the user}

\item{token}{character, api token}

\item{max_retries}{integer, maximum number of retries on failure}

\item{retry_delay}{numeric, delay in seconds between retries}
}
\value{
list of all follow records with uri, cid, and value fields
}
\description{
Get all follow records from user's repository
}
