% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_follows}
\alias{get_follows}
\title{Get follows for an actor}
\usage{
get_follows(actor, token)
}
\arguments{
\item{actor}{account identifier}

\item{token}{token for api}
}
\value{
tibble
}
\description{
Get follows for an actor
}
\section{Lexicon references}{

\href{https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/graph/getFollows.json}{lexicons/app/bsky/graph/getFollows.json}
}

