% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_token}
\alias{get_token}
\title{Get token for Bluesky api}
\usage{
get_token(identifier, password)
}
\arguments{
\item{identifier}{identifier for account}

\item{password}{app password}
}
\value{
response body
}
\description{
Get token for Bluesky api
}
