% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algos.R
\name{init_net}
\alias{init_net}
\title{Create initial network}
\usage{
init_net(actor, keywords, token)
}
\arguments{
\item{actor}{account identifier}

\item{keywords}{character vector with keywords that we check for in actors' description}

\item{token}{token for api}
}
\value{
tibble with two columns, two columns: \code{actor_handle}
(the Blue Sky Social actor who is followING another) and \code{follows_handle}
(the actor being followed).
}
\description{
Get all follows for a single actor, and use that to
form a minimal, initial network that can be expanded
later
}
