% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{refresh_token}
\alias{refresh_token}
\title{Refresh token for Bluesky api}
\usage{
refresh_token(refresh_token)
}
\arguments{
\item{refresh_token}{character, refresh token}
}
\value{
response body
}
\description{
Refresh token for Bluesky api
}
