% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{word_freqs}
\alias{word_freqs}
\title{Word frequencies}
\usage{
word_freqs(
  texts,
  bigrams = FALSE,
  top = 30,
  remove_stopwords = TRUE,
  language = "en"
)
}
\arguments{
\item{texts}{character. Vector of texts (like profile descriptions)}

\item{bigrams}{boolean. Should we count bigrams? Default to FALSE}

\item{top}{integer. The top words/bi-grams to include. Defaults to 30}

\item{remove_stopwords}{boolean. Should stopwords be excluded? Defaults to TRUE}

\item{language}{character. The language to use when removing stopwords. Defaults to "en"}
}
\value{
dataframe with word (or bi-gram) frequencies
}
\description{
Word frequencies
}
